requirements.txt
pyYAML >= 3.12
